Updated packaged and updated the DiscordClientWrapper
