+++ package.json
@@ -11,12 +11,6 @@
         "sql-test": "heroku pg:psql -a pubg-stat-discord-bot-dev"
     },
     "dependencies": {
-        "@types/cheerio": "^0.22.7",
-        "@types/es6-promise": "^3.3.0",
-        "@types/node": "^10.3.0",
-        "@types/pg": "^7.4.9",
-        "@types/request-promise": "^4.1.41",
-        "@types/winston": "^2.3.9",
         "cheerio": "latest",
         "curlrequest": "^1.0.1",
         "discord.js": "^11.3.0",
@@ -24,10 +18,16 @@
         "pg": "^7.4.1",
         "request": "^2.83.0",
         "request-promise": "^4.2.2",
-        "typescript": "^2.9.1",
         "winston": "^2.4.0"
     },
     "devDependencies": {
-        "eslint": "^4.19.1"
+        "@types/cheerio": "^0.22.7",
+        "@types/es6-promise": "^3.3.0",
+        "@types/node": "^10.3.0",
+        "@types/pg": "^7.4.9",
+        "@types/request-promise": "^4.1.41",
+        "@types/winston": "^2.3.9",
+        "eslint": "^4.19.1",
+        "typescript": "^2.9.1"
     }
 }
