--- src/discord-client.ts
@@ -1,10 +1,0 @@
-import * as Discord from 'discord.js';
-
-
-export class DiscordClientWrapper extends Discord.Client {
-    aliases: any;
-    commands: any;
-    channel: any;
-    elevation(arg0: any): any {};
-    reload: (command: any) => Promise<{}>;
-}
